ref: simplify and move observation dict def
