+++ gym_multigrid/policy/ctf/typing.py
@@ -1,0 +1,15 @@
+from typing import TypedDict
+
+import numpy as np
+from numpy.typing import NDArray
+
+
+class ObservationDict(TypedDict):
+    blue_agent: NDArray[np.int_]
+    red_agent: NDArray[np.int_]
+    blue_flag: NDArray[np.int_]
+    red_flag: NDArray[np.int_]
+    blue_territory: NDArray[np.int_]
+    red_territory: NDArray[np.int_]
+    obstacle: NDArray[np.int_]
+    terminated_agents: NDArray[np.int_]
